factored out helpers funcs
